requirements.txt
absl-py==1.4.0
accelerate==0.19.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
astor==0.8.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
brotlipy==0.7.0
cachetools==5.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.0.7
croniter==1.3.14
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.3.0
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
einops==0.6.1
et-xmlfile==1.1.0
fastapi==0.88.0
filelock @ file:///croot/filelock_1672387128942/work
flake8==6.0.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455532332/work
google-auth==2.18.0
google-auth-oauthlib==1.0.0
gpustat==1.1
grpcio==1.54.0
h11==0.14.0
huggingface-hub==0.14.1
idna @ file:///croot/idna_1666125576474/work
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.2.0
json-tricks==3.17.1
kiwisolver==1.4.4
lazy_loader==0.2
lightning==2.0.2
lightning-cloud==0.5.34
lightning-utilities==0.8.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.1
mccabe==0.7.0
mdurl==0.1.2
MedPy==0.4.0
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
mmcv==2.0.0
mmengine==0.7.3
model-index==0.1.11
monai==1.1.0
mpmath==1.2.1
multidict==6.0.4
munch==2.5.0
networkx @ file:///croot/networkx_1678964333703/work
nni==2.10.1
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
nvidia-ml-py==11.525.112
oauthlib==3.2.2
opencv-python==4.7.0.72
openmim==0.3.7
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
pathtools==0.1.2
Pillow==9.4.0
pretrainedmodels==0.7.4
prettytable==3.7.0
protobuf==3.20.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.6.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
PythonWebHDFS==0.2.3
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
readchar==4.0.5
regex==2023.5.5
requests @ file:///croot/requests_1682607517574/work
requests-oauthlib==1.3.1
responses==0.23.1
rich==13.3.5
rsa==4.9
schema==0.7.5
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
segmentation-models-pytorch==0.3.2
sentry-sdk==1.26.0
setproctitle==1.3.2
SimpleITK==2.2.1
simplejson==3.19.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.22.0
starsessions==1.3.0
sympy @ file:///croot/sympy_1668202399572/work
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.6
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.6.12
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.0
torch-tb-profiler==0.4.1
torchaudio==2.0.0
torchmetrics==0.11.4
torchvision==0.15.0
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.0
triton==2.0.0
typeguard==2.13.3
types-PyYAML==6.0.12.10
typing_extensions @ file:///croot/typing_extensions_1681939499988/work
tzdata==2023.3
urllib3 @ file:///croot/urllib3_1680254681959/work
uvicorn==0.22.0
wandb==0.15.4
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.3.4
yapf==0.33.0
yarl==1.9.2
zipp==3.15.0
